libobs-d3d11,libobs-opengl,plugins: Remove unneeded cast in switches
Those casts were usually used to workaround switch warnings.

+++ plugins/win-capture/graphics-hook/dxgi-helpers.hpp
@@ -4,7 +4,7 @@
 
 static inline DXGI_FORMAT strip_dxgi_format_srgb(DXGI_FORMAT format)
 {
-	switch ((unsigned long)format) {
+	switch (format) {
 	case DXGI_FORMAT_B8G8R8A8_UNORM_SRGB:
 		return DXGI_FORMAT_B8G8R8A8_UNORM;
 	case DXGI_FORMAT_R8G8B8A8_UNORM_SRGB:
@@ -18,7 +18,7 @@ static inline DXGI_FORMAT apply_dxgi_format_typeless(DXGI_FORMAT format,
 						     bool allow_srgb_alias)
 {
 	if (allow_srgb_alias) {
-		switch ((unsigned long)format) {
+		switch (format) {
 		case DXGI_FORMAT_B8G8R8A8_UNORM:
 			return DXGI_FORMAT_B8G8R8A8_TYPELESS;
 		case DXGI_FORMAT_R8G8B8A8_UNORM:
